Revert "Shit"
This reverts commit 807046f1ee4516796e4d709c5bad8e6f357f33d5.

+++ MathModule/Actions/BMF.cs
@@ -8,9 +8,9 @@ namespace MathModule
 {
     public class BMF
     {
-        public static decimal epsilon = 1E-7m;
+        public static double epsilon = 1E-7;
 
-        public static bool Deq(decimal d1, decimal d2)
+        public static bool Deq(double d1, double d2)
         {
             if (d1 == d2)
             {
@@ -18,7 +18,7 @@ namespace MathModule
             }
             return Math.Abs(d1 - d2) <= epsilon;
         }
-        public static bool Deq(decimal d1, decimal d2, decimal epsilon)
+        public static bool Deq(double d1, double d2, double epsilon)
         {
             if (d1 == d2)
             {
@@ -26,22 +26,22 @@ namespace MathModule
             }
             return Math.Abs(d1 - d2) <= epsilon;
         }
-        public static List<Point> Interpol(TriangulationEdge edge, decimal dH)
+        public static List<Point> Interpol(TriangulationEdge edge, double dH)
         {
             List<Point> result = new List<Point>();
             Point p1 = edge.points[0];
             Point p2 = edge.points[1];
-            decimal height1 = p1.Z;
-            decimal height2 = p2.Z;
-            decimal dheight = Math.Abs(height1 - height2);
-            decimal diff = height1 % dH;
-            decimal length = Math.Sqrt(Math.Pow(p2.X - p1.X, 2) + Math.Pow(p2.Y - p1.Y, 2));
-            decimal ang = AngleByCoordinates(p1, p2);
-            decimal tan = (p1.Z - p2.Z) / length;
-            decimal x;
-            decimal y;
-            decimal z;
-            decimal dist = dH / dheight * length;
+            double height1 = p1.Z;
+            double height2 = p2.Z;
+            double dheight = Math.Abs(height1 - height2);
+            double diff = height1 % dH;
+            double length = Math.Sqrt(Math.Pow(p2.X - p1.X, 2) + Math.Pow(p2.Y - p1.Y, 2));
+            double ang = AngleByCoordinates(p1, p2);
+            double tan = (p1.Z - p2.Z) / length;
+            double x;
+            double y;
+            double z;
+            double dist = dH / dheight * length;
             // Придумать название переменной, или использовать усложненные формулы.
             // Проверка на присутствие горизнтали между точками
             if (dheight < dH)
@@ -51,7 +51,7 @@ namespace MathModule
                     return null;
                 }
             }
-            if (Deq(diff, 0m))
+            if (Deq(diff, 0d))
             {
                 diff = dH;
             }
@@ -61,7 +61,7 @@ namespace MathModule
                 diff = height2 < height1 ? diff : dH - diff;
             }
             // Основной цикл отрисовки. Требуется улучшить воспринимаемость.
-            for (decimal i = dist * diff / dH; Math.Round(i, 4) < length; i += dist)
+            for (double i = dist * diff / dH; Math.Round(i, 4) < length; i += dist)
             {
                 // "Вверх"
                 if (Deq(p1.X, p2.X) && p2.Y - p1.Y > 0)
@@ -103,25 +103,25 @@ namespace MathModule
         }
 
         // При интерполировании считается, что Х - вертикально направленная ось
-        public static Point InterpolToPoint(TriangulationEdge edge, decimal PointHeight)
+        public static Point InterpolToPoint(TriangulationEdge edge, double PointHeight)
         {
             // Точки.
             Point p1 = edge.points[0];
             Point p2 = edge.points[1];
             // Высоты точек.
-            decimal height1 = p1.Z;
-            decimal height2 = p2.Z;
+            double height1 = p1.Z;
+            double height2 = p2.Z;
             // Превышение между точками.
-            decimal dheight = Math.Abs(height1 - height2);
+            double dheight = Math.Abs(height1 - height2);
             // Превышение между первой и искомой точкой.
-            decimal diff = height1 > height2 ? Math.Abs(height1 - (PointHeight % height1))   :  PointHeight - (height1 % PointHeight);
-            decimal length = Math.Sqrt(Math.Pow(p2.X - p1.X, 2) + Math.Pow(p2.Y - p1.Y, 2));
-            decimal ang = AngleByCoordinates_Clockwise(p1, p2);
-            decimal tan = (height2 - height1) / length;
-            decimal x;
-            decimal y;
-            decimal z;
-            decimal dist = diff / dheight * length;
+            double diff = height1 > height2 ? Math.Abs(height1 - (PointHeight % height1))   :  PointHeight - (height1 % PointHeight);
+            double length = Math.Sqrt(Math.Pow(p2.X - p1.X, 2) + Math.Pow(p2.Y - p1.Y, 2));
+            double ang = AngleByCoordinates_Clockwise(p1, p2);
+            double tan = (height2 - height1) / length;
+            double x;
+            double y;
+            double z;
+            double dist = diff / dheight * length;
             // Придумать название переменной, или использовать усложненные формулы.
             // Проверка на присутствие горизнтали между точками
             if(edge.points[0].Z > PointHeight && edge.points[1].Z > PointHeight || edge.points[0].Z < PointHeight && edge.points[1].Z < PointHeight)
@@ -129,28 +129,28 @@ namespace MathModule
                 return null;
             }
             // "Вверх"
-            if (Deq(p1.X, p2.X) && p2.Y - p1.Y > 0m)
+            if (Deq(p1.X, p2.X) && p2.Y - p1.Y > 0d)
             {
                 x = p1.X;
                 y = p1.Y + dist;
                 z = p1.Z + dist * tan;
             }
             // "Вниз"
-            else if (Deq(p1.X, p2.X) && p2.Y - p1.Y < 0m)
+            else if (Deq(p1.X, p2.X) && p2.Y - p1.Y < 0d)
             {
                 x = p1.X;
                 y = p1.Y - dist;
                 z = p1.Z + dist * tan;
             }
             // "Вправо"
-            else if (Deq(p1.Y, p2.Y) && p2.X - p1.X > 0m)
+            else if (Deq(p1.Y, p2.Y) && p2.X - p1.X > 0d)
             {
                 x = p1.X + dist;
                 y = p1.Y;
                 z = p1.Z + dist * tan;
             }
             // "Влево"
-            else if (Deq(p1.Y, p2.Y) && p2.X - p1.X < 0m)
+            else if (Deq(p1.Y, p2.Y) && p2.X - p1.X < 0d)
             {
                 x = p1.X - dist;
                 y = p1.Y;
@@ -167,13 +167,13 @@ namespace MathModule
 
 
         // Служебно-математические функции, нужны для вычислений.
-        public static decimal[] CoordinatesByAngle(Point origin, decimal angle, decimal distance)
+        public static double[] CoordinatesByAngle(Point origin, double angle, double distance)
         {
-            decimal[] resultXY = new decimal[2];
+            double[] resultXY = new double[2];
 
             switch (angle)
             {
-                case 0m:
+                case 0d:
                     resultXY[0] = origin.X;
                     resultXY[1] = origin.Y + distance;
                     return resultXY;
@@ -197,55 +197,55 @@ namespace MathModule
 
         /// <summary>
         ///  Крайне полезная и, в будущем, скорее всего, основная функция расчета углов из координат. Работает, не трогай. На вход требует
-        /// 2 Point и выдаёт дирекционный угол в виде decimal вторая и четвёртая четверть перевёрнуты из-за того, что в автокаде Y идёт в обратную сторону от прямоугольной 
+        /// 2 Point и выдаёт дирекционный угол в виде double вторая и четвёртая четверть перевёрнуты из-за того, что в автокаде Y идёт в обратную сторону от прямоугольной 
         /// системы координат.
         /// </summary>
         /// <param name="p1"></param>
         /// <param name="p2"></param>
         /// <returns></returns>
-        public static decimal AngleByCoordinates_Clockwise(Point p1, Point p2)
+        public static double AngleByCoordinates_Clockwise(Point p1, Point p2)
         {
-            decimal dX = p2.X - p1.X;
-            decimal dY = p2.Y - p1.Y;
-            if (Deq(dX, 0m) && Deq(dY, 0m))
+            double dX = p2.X - p1.X;
+            double dY = p2.Y - p1.Y;
+            if (Deq(dX, 0d) && Deq(dY, 0d))
             {
                 return 0;
             }
-            if (Deq(dX, 0m) && dY > 0m)
+            if (Deq(dX, 0d) && dY > 0d)
             {
                 return Math.PI / 2d;
             }
-            if (Deq(dX, 0m) && dY < 0m)
+            if (Deq(dX, 0d) && dY < 0d)
             {
                 return Math.PI * 1.5d;
             }
-            if (dX > 0 && Deq(dY, 0m))
+            if (dX > 0 && Deq(dY, 0d))
             {
-                return 0m;
+                return 0d;
             }
-            if (dX < 0 && Deq(dY, 0m))
+            if (dX < 0 && Deq(dY, 0d))
             {
                 return Math.PI;
             }
-            decimal rum = Math.Atan(Math.Abs(dY / dX));
-            decimal ang = 0m;
+            double rum = Math.Atan(Math.Abs(dY / dX));
+            double ang = 0d;
             // Первая четверть
-            if (dX > 0m && dY > 0m)
+            if (dX > 0d && dY > 0d)
             {
                 ang = rum;
             }
             // Вторая четверть
-            else if (dX < 0m && dY > 0m)
+            else if (dX < 0d && dY > 0d)
             {
                 ang =  Math.PI - rum;
             }
             //Третья четверть
-            else if (dX < 0m && dY < 0m)
+            else if (dX < 0d && dY < 0d)
             {
                 ang = Math.PI + rum;
             }
             //Четвертая четверть
-            else if (dX > 0m && dY < 0m)
+            else if (dX > 0d && dY < 0d)
             {
 
                 ang = 2d * Math.PI - rum;
@@ -257,56 +257,56 @@ namespace MathModule
 
         /// <summary>
         ///  Крайне полезная и, в будущем, скорее всего, основная функция расчета углов из координат. Работает, не трогай. На вход требует
-        /// 2 Point и выдаёт дирекционный угол в виде decimal вторая и четвёртая четверть перевёрнуты из-за того, что в автокаде Y идёт в обратную сторону от прямоугольной 
+        /// 2 Point и выдаёт дирекционный угол в виде double вторая и четвёртая четверть перевёрнуты из-за того, что в автокаде Y идёт в обратную сторону от прямоугольной 
         /// системы координат.
         /// </summary>
         /// <param name="p1"></param>
         /// <param name="p2"></param>
         /// <returns></returns>
-        public static decimal AngleByCoordinates(Point p1, Point p2)
+        public static double AngleByCoordinates(Point p1, Point p2)
         {
-            decimal dX = p2.X - p1.X;
-            decimal dY = p2.Y - p1.Y;
-            if(Deq(dX,0m) && Deq(dY, 0m))
+            double dX = p2.X - p1.X;
+            double dY = p2.Y - p1.Y;
+            if(Deq(dX,0d) && Deq(dY, 0d))
             {
-                return 0m;
+                return 0d;
             }
-            if (Deq(dX, 0m) && dY > 0m)
+            if (Deq(dX, 0d) && dY > 0d)
             {
                 return Math.PI + Math.PI / 2d;
             }
-            if (Deq(dX, 0m) && dY < 0m)
+            if (Deq(dX, 0d) && dY < 0d)
             {
                 return Math.PI / 2d;
             }
-            if(dX > 0 && Deq(dY, 0m))
+            if(dX > 0 && Deq(dY, 0d))
             {
-                return 0m;
+                return 0d;
             }
-            if (dX < 0 && Deq(dY, 0m))
+            if (dX < 0 && Deq(dY, 0d))
             {
                 return Math.PI;
             }
-            decimal rum = Math.Atan(Math.Abs(dY / dX));
-            decimal ang = 0m;
+            double rum = Math.Atan(Math.Abs(dY / dX));
+            double ang = 0d;
             // Первая четверть (4)
-            if (dX > 0m && dY > 0m)
+            if (dX > 0d && dY > 0d)
             {
                 ang = 2d * Math.PI - rum;
             }
             // Вторая четверть (3)
-            else if (dX < 0m && dY > 0m)
+            else if (dX < 0d && dY > 0d)
             {
                 ang = Math.PI + rum;
 
             }
             //Третья четверть (2)
-            else if (dX < 0m && dY < 0m)
+            else if (dX < 0d && dY < 0d)
             {
                 ang = Math.PI - rum;
             }
             //Четвертая четверть (1)
-            else if (dX > 0m && dY < 0m)
+            else if (dX > 0d && dY < 0d)
             {
                 ang = rum;
             }
@@ -320,7 +320,7 @@ namespace MathModule
         /// <param name="p1"></param>
         /// <param name="p2"></param>
         /// <returns></returns>
-        public static decimal GetDistance2D(Point p1, Point p2)
+        public static double GetDistance2D(Point p1, Point p2)
         {
             return Math.Sqrt(Math.Pow((p1.X - p2.X),2) + Math.Pow((p1.Y - p2.Y),2));
         }
@@ -332,11 +332,11 @@ namespace MathModule
         /// <param name="segment"></param>
         /// <param name="point"></param>
         /// <returns></returns>
-        public static decimal SideRelativelyOfSegment(Segment segment, Point point)
+        public static double SideRelativelyOfSegment(Segment segment, Point point)
         {
             return (segment.points[0].X - point.X) * (segment.points[1].Y - point.Y) - (segment.points[1].X - point.X) * (segment.points[0].Y - point.Y);
         }
-        public static decimal DistanceBetweenPointAndPlane(Point point, decimal[] planeFactors)
+        public static double DistanceBetweenPointAndPlane(Point point, double[] planeFactors)
         {
             return (point.X * planeFactors[0] + point.Y * planeFactors[1] + point.Z * planeFactors[2] + planeFactors[3]) /
                 Math.Sqrt(Math.Pow(planeFactors[0], 2) + Math.Pow(planeFactors[1], 2) + Math.Pow(planeFactors[2], 2));
@@ -344,9 +344,9 @@ namespace MathModule
 
         public static Point GetCenter(List<Point> points)
         {
-            decimal x = 0;
-            decimal y = 0;
-            decimal z = 0;
+            double x = 0;
+            double y = 0;
+            double z = 0;
             foreach(Point point in points)
             {
                 x += point.X;
@@ -364,9 +364,9 @@ namespace MathModule
         /// <returns></returns>
         public static Point GetMin(List<Point> points)
         {
-            decimal xMin = decimal.MaxValue;
-            decimal yMin = decimal.MaxValue;
-            decimal zMin = decimal.MaxValue;
+            double xMin = double.MaxValue;
+            double yMin = double.MaxValue;
+            double zMin = double.MaxValue;
 
             foreach (Point point in points)
             {
@@ -392,9 +392,9 @@ namespace MathModule
         /// <returns></returns>
         public static Point GetMax(List<Point> points)
         {
-            decimal xMax = decimal.MinValue;
-            decimal yMax = decimal.MinValue;
-            decimal zMax = decimal.MinValue;
+            double xMax = double.MinValue;
+            double yMax = double.MinValue;
+            double zMax = double.MinValue;
 
             foreach (Point point in points)
             {
@@ -431,8 +431,8 @@ namespace MathModule
         }
         public static bool CompatePointsByAngle(Point basis, Point p1, Point p2)
         {
-            decimal ang1 = BMF.AngleByCoordinates_Clockwise(basis, p1);
-            decimal ang2 = BMF.AngleByCoordinates_Clockwise(basis, p2);
+            double ang1 = BMF.AngleByCoordinates_Clockwise(basis, p1);
+            double ang2 = BMF.AngleByCoordinates_Clockwise(basis, p2);
             if (ang1 > ang2)
             {
                 return true;
