+++ MathModule/Primitives/ExperimentsAndRework/Polyline.cs
@@ -128,21 +128,21 @@ namespace MathModule
         {
             this.segments.Insert(0, new Segment(point, this.segments[0].points[0]));
         }
-        public decimal GetArea()
+        public double GetArea()
         {
-            decimal result = 0m;
+            double result = 0d;
             for (int i = 0; i < GetPoints().Count - 1; i++)
             {
-                result += MxF.Det(new decimal[2, 2] { { GetPoints()[i].X, GetPoints()[i].Y }, { GetPoints()[i + 1].X, GetPoints()[i + 1].Y } });
+                result += MxF.Det(new double[2, 2] { { GetPoints()[i].X, GetPoints()[i].Y }, { GetPoints()[i + 1].X, GetPoints()[i + 1].Y } });
             }
-            result += MxF.Det(new decimal[2, 2] { { GetPoints()[GetPoints().Count - 1].X, GetPoints()[GetPoints().Count - 1].Y }, { GetPoints()[0].X, GetPoints()[0].Y } });
+            result += MxF.Det(new double[2, 2] { { GetPoints()[GetPoints().Count - 1].X, GetPoints()[GetPoints().Count - 1].Y }, { GetPoints()[0].X, GetPoints()[0].Y } });
             return Math.Abs(result / 2);
         }
         public Point GetMin()
         {
-            decimal xMin = decimal.MaxValue;
-            decimal yMin = decimal.MaxValue;
-            decimal zMin = decimal.MaxValue;
+            double xMin = double.MaxValue;
+            double yMin = double.MaxValue;
+            double zMin = double.MaxValue;
 
             foreach (Point point in this.GetPoints())
             {
@@ -163,9 +163,9 @@ namespace MathModule
         }
         public Point GetMax()
         {
-            decimal xMax = decimal.MinValue;
-            decimal yMax = decimal.MinValue;
-            decimal zMax = decimal.MinValue;
+            double xMax = double.MinValue;
+            double yMax = double.MinValue;
+            double zMax = double.MinValue;
 
             foreach (Point point in this.GetPoints())
             {
@@ -187,9 +187,9 @@ namespace MathModule
         public Point GetCenter()
         {
             List<Point> points = this.GetPoints();
-            decimal X = 0;
-            decimal Y = 0;
-            decimal Z = 0;
+            double X = 0;
+            double Y = 0;
+            double Z = 0;
             foreach (Point point in points)
             {
                 X += point.X;
