+++ MathModule/TriangulationNet/Triangulation.cs
@@ -573,19 +573,19 @@ namespace MathModule
         private bool IsDelone(Triangle tr1, Triangle tr2)
         {
             Point[] points = tr1.GetPoints(tr2);
-            decimal dx0 = (points[0].X - points[1].X);
-            decimal dx1 = (points[0].X - points[3].X);
-            decimal dy0 = (points[0].Y - points[1].Y);
-            decimal dy1 = (points[0].Y - points[3].Y);
+            double dx0 = (points[0].X - points[1].X);
+            double dx1 = (points[0].X - points[3].X);
+            double dy0 = (points[0].Y - points[1].Y);
+            double dy1 = (points[0].Y - points[3].Y);
 
-            decimal dx2 = (points[2].X - points[1].X);
-            decimal dx3 = (points[2].X - points[3].X);
-            decimal dy2 = (points[2].Y - points[1].Y);
-            decimal dy3 = (points[2].Y - points[3].Y);
+            double dx2 = (points[2].X - points[1].X);
+            double dx3 = (points[2].X - points[3].X);
+            double dy2 = (points[2].Y - points[1].Y);
+            double dy3 = (points[2].Y - points[3].Y);
 
-            decimal st1 = dx0 * dx1 + dy0 * dy1;
-            decimal st2 = dx2 * dx3 + dy2 * dy3;
-            if (st1 >= 0m && st2 >= 0m)
+            double st1 = dx0 * dx1 + dy0 * dy1;
+            double st2 = dx2 * dx3 + dy2 * dy3;
+            if (st1 >= 0d && st2 >= 0d)
             {
                 return true;
             }
@@ -593,7 +593,7 @@ namespace MathModule
             return (Math.Abs(dx0 * dy1 - dx1 * dy0)
                 * (st2)
                 + (st1)
-                * Math.Abs(dx3 * dy2 - dx2 * dy3)) >= 0m;
+                * Math.Abs(dx3 * dy2 - dx2 * dy3)) >= 0d;
         }
 
 
@@ -606,7 +606,7 @@ namespace MathModule
         /// <returns></returns>
         private bool IsCollinear(Point p1, Point p2, Point p3)
         {
-            if (BMF.Deq((p1.X - p3.X) * (p2.Y - p3.Y) - (p2.X - p3.X) * (p1.Y - p3.Y), 0m))
+            if (BMF.Deq((p1.X - p3.X) * (p2.Y - p3.Y) - (p2.X - p3.X) * (p1.Y - p3.Y), 0d))
             {
                 return true;
             }
@@ -1109,7 +1109,7 @@ namespace MathModule
 
 
 
-        public void CreateIsolines(decimal step)
+        public void CreateIsolines(double step)
         {
             try
             {
@@ -1119,8 +1119,8 @@ namespace MathModule
                 Point min = searchArray.GetMin();
                 Point max = searchArray.GetMax();
 
-                decimal absoluteHeight = min.Z + (step - (min.Z % step));
-                decimal currentHeight = absoluteHeight;
+                double absoluteHeight = min.Z + (step - (min.Z % step));
+                double currentHeight = absoluteHeight;
                 bool isEdgeOnPlaneHeight = false;
                 while (absoluteHeight < max.Z)
                 {
@@ -1154,7 +1154,7 @@ namespace MathModule
                 MessageBox.Show(e.ToString());
             }
         }
-        public void CreateBorderIsoline(decimal height)
+        public void CreateBorderIsoline(double height)
         {
             foreach (TriangulationEdge edge in outerEdges)
             {
@@ -1235,7 +1235,7 @@ namespace MathModule
 
 
 
-        public void CreateInnerIsoline(decimal height)
+        public void CreateInnerIsoline(double height)
         {
             foreach (Triangle tr in triangles)
             {
@@ -1284,7 +1284,7 @@ namespace MathModule
                 innerIsolines.Add(isoline);
             }
         }
-        private void MarkTrianglesForIsoline(decimal height)
+        private void MarkTrianglesForIsoline(double height)
         {
             foreach (Triangle tr in triangles)
             {
@@ -1298,7 +1298,7 @@ namespace MathModule
                 }
             }
         }
-        private bool IsHeightInEdge(TriangulationEdge edge, decimal height)
+        private bool IsHeightInEdge(TriangulationEdge edge, double height)
         {
             if ((edge.points[0].Z <= height && height <= edge.points[1].Z) || (edge.points[0].Z >= height && height >= edge.points[1].Z))
             {
